--- src/lab
@@ -1,5 +1,0 @@
-11111
-1s1f1
-10101
-10001
-11111
